requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
unstructured
google-generativeai
